// src/app/model/api/Arbeitslosenkasse.ts
import {AbstractApiElement} from './AbstractApiElement';

export class Arbeitslosenkasse extends AbstractApiElement {
  id: number = null;
  kantonId: number = null;
  kassenNr: string = null;
  adresse: string = null;
  name: string = null;
  tel: string = null;
  email: string = null;
  fax: string = null;
  link: string = null;
}